+++ src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import App from './app.vue';
-import store from './store.js';
+import createStore from './store';
 import dayjs from 'dayjs';
 import 'dayjs/locale/de';
 import localizedFormat from 'dayjs/plugin/localizedFormat';
@@ -8,8 +8,12 @@ import localizedFormat from 'dayjs/plugin/localizedFormat';
 dayjs.locale('de');
 dayjs.extend(localizedFormat);
 
-export default new Vue({
-	store,
-	el: '#content',
-	render: h => h(App),
-});
+(async() => {
+	const store = await createStore();
+
+	return new Vue({
+		store,
+		el: '#content',
+		render: h => h(App),
+	});
+})();